Create README.md
Modules and hours allocated

The user will be able to enter the modules he/she is doing. The user will also then have to enter the module details such as code, name, number of credits as well as the amount of hours per week. The program calculates the amount of hours of self study needed to be completed, with the calculation being num of credits x 10 divided by the number of weeks, then the class hours per week will be subtracted from that total.

a. How to compile and run software
-	Download visio studio 2019
-	After download, select and install workloads
-	Launch software
-	Create new project(wpf)
-	Copy the source code
-	Paste source code in your new project
-	Run file